fix false price triggering bug

+++ src/main/kotlin/io/ambershogun/mentatus/notification/price/PriceNotificationRepository.kt
@@ -9,8 +9,8 @@ interface PriceNotificationRepository : MongoRepository<PriceNotification, Strin
     fun deleteAllByChatId(chatId: Long)
 
     @Query(value = "{ 'ticker' : ?0, 'equitySign':  ?1, 'price':  {'\$gte': ?2}}")
-    fun findByTickerAndPriceGreaterThen(ticker: String, equitySign: EquitySign, price: BigDecimal): List<PriceNotification>
+    fun findByTickerAndPriceGreaterThen(ticker: String, equitySign: EquitySign, price: Double): List<PriceNotification>
 
     @Query(value = "{ 'ticker' : ?0, 'equitySign':  ?1, 'price':  {'\$lte': ?2}}")
-    fun findByTickerAndPriceLessThen(ticker: String, equitySign: EquitySign, price: BigDecimal): List<PriceNotification>
+    fun findByTickerAndPriceLessThen(ticker: String, equitySign: EquitySign, price: Double): List<PriceNotification>
 }
